dont overwrite uoproxy.conf
git-svn-id: svn://svn.berlios.de/uoproxy/trunk@468 546381f1-97f9-0310-94fd-acf1737c2c57

+++ Makefile
@@ -59,7 +59,7 @@ src/uoproxy: $(OBJECTS)
 
 install: src/uoproxy
 	install -d -m 0755 /etc $(PREFIX)/bin
-	install -m 0644 conf/uoproxy.conf /etc
+	test -f /etc/uoproxy.conf || install -m 0644 conf/uoproxy.conf /etc
 	install -m 0755 src/uoproxy $(PREFIX)/bin
 
 strip: src/uoproxy
